Várias mudanças, adiçoes e melhorias.
