+++ src/main/kotlin/br/com/api/kotlin/Imports.kt
@@ -1,0 +1,10 @@
+package br.com.api.kotlin
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class EstudarApp
+
+fun main(args: Array<String>) {
+    runApplication<br.com.api.kotlin.EstudarApp>(*args)
+}
