+++ src/main/kotlin/br/com/api/kotlin/entity/Category.kt
@@ -1,0 +1,21 @@
+package br.com.api.kotlin.entity
+
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+import javax.persistence.Table
+
+@Entity
+@Table(name = "CATEGORY")
+class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null
+
+    @Column
+    var category: String? = null
+
+}
