+++ src/main/kotlin/br/com/api/kotlin/enumerator/PaymentType.kt
@@ -1,0 +1,10 @@
+package br.com.api.kotlin.enumerator
+
+enum class PaymentType {
+
+    DEBITO,
+    PIX,
+    AVISTA,
+    APRAZO
+
+}
