+++ src/main/kotlin/br/com/api/kotlin/repository/ClientRepository.kt
@@ -1,0 +1,14 @@
+package br.com.api.kotlin.repository
+
+import br.com.api.kotlin.entity.Client
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Query
+import java.util.*
+
+interface ClientRepository : JpaRepository<br.com.api.kotlin.entity.Client, Long> {
+
+
+    @Query("SELECT cl FROM Client cl WHERE cl.cpf = :cpf")
+    fun findByCpf(cpf: String?): Optional<br.com.api.kotlin.entity.Client>
+
+}
