+++ src/main/kotlin/br/com/api/kotlin/repository/CpuRepository.kt
@@ -1,0 +1,8 @@
+package br.com.api.kotlin.repository
+
+import br.com.api.kotlin.entity.Cpu
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CpuRepository: JpaRepository<br.com.api.kotlin.entity.Cpu,Long>{
+
+}
