+++ src/main/kotlin/br/com/api/kotlin/repository/RamRepository.kt
@@ -1,0 +1,7 @@
+package br.com.api.kotlin.repository
+
+import br.com.api.kotlin.entity.Ram
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface RamRepository: JpaRepository<br.com.api.kotlin.entity.Ram, Long> {
+}
