--- src/main/kotlin/com/example/demo/dto/CpuDto.kt
@@ -1,4 +1,0 @@
-package com.example.demo.dto
-
-class CpuDto {
-}
