--- src/main/kotlin/com/example/demo/repository/AddressRepository.kt
@@ -1,7 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.entity.Address
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface AddressRepository : JpaRepository<Address, Long> {
-}
