--- src/main/kotlin/com/example/demo/repository/CpuRepository.kt
@@ -1,8 +1,0 @@
-package com.example.demo.repository
-
-import com.example.demo.entity.Cpu
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CpuRepository: JpaRepository<Cpu,Long>{
-
-}
