--- src/main/kotlin/com/example/demo/repository/RamRepository.kt
@@ -1,7 +1,0 @@
-package com.example.demo.repository
-
-import com.example.demo.entity.Ram
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface RamRepository: JpaRepository<Ram, Long> {
-}
