--- src/main/kotlin/com/example/demo/repository/SsdRepository.kt
@@ -1,13 +1,0 @@
-package com.example.demo.repository
-
-import com.example.demo.entity.Ssd
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.Query
-
-interface SsdRepository: JpaRepository<Ssd, Long> {
-
-    @Query("SELECT p.id FROM Ssd p WHERE purchase_date = arrival_date ")
-    fun checaTeste(): List<Ssd>
-
-
-}
