+++ src/test/kotlin/br/com/api/kotlin/EstudarApplicationTests.kt
@@ -1,0 +1,13 @@
+package br.com.api.kotlin
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EstudarApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
