--- src/test/kotlin/com/example/demo/EstudarApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.demo
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class EstudarApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
